+)添加xss过滤 +)用户管理 *)优化依赖

+++ zero-boot/zero-boot-system/src/main/java/com/fast/system/controller/SysUserController.java
@@ -1,18 +1,26 @@
 package com.fast.system.controller;
 
+import cn.hutool.core.util.StrUtil;
 import com.fast.framework.common.bean.PageResult;
 import com.fast.framework.common.bean.Result;
+import com.fast.framework.security.user.SecurityUser;
+import com.fast.framework.security.user.UserDetail;
 import com.fast.system.convert.SysUserConvert;
 import com.fast.system.entity.SysUserEntity;
 import com.fast.system.query.SysUserQuery;
+import com.fast.system.service.SysUserPostService;
+import com.fast.system.service.SysUserRoleService;
 import com.fast.system.service.SysUserService;
+import com.fast.system.vo.SysUserPasswordVO;
 import com.fast.system.vo.SysUserVO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
 
@@ -28,6 +36,9 @@ import java.util.List;
 @AllArgsConstructor
 public class SysUserController {
     private final SysUserService sysUserService;
+    private final SysUserRoleService sysUserRoleService;
+    private final SysUserPostService sysUserPostService;
+    private final PasswordEncoder passwordEncoder;
 
     @GetMapping("page" )
     @Operation(summary = "分页" )
@@ -43,14 +54,55 @@ public class SysUserController {
     @PreAuthorize("hasAuthority('sys:user:info')" )
     public Result<SysUserVO> get(@PathVariable("id" ) Long id) {
         SysUserEntity entity = sysUserService.getById(id);
+        SysUserVO vo = SysUserConvert.INSTANCE.convert(entity);
 
-        return Result.ok(SysUserConvert.INSTANCE.convert(entity));
+        // 用户角色列表
+        List<Long> roleIdList = sysUserRoleService.getRoleIdList(id);
+        vo.setRoleIdList(roleIdList);
+
+        // 用户岗位列表
+        List<Long> postIdList = sysUserPostService.getPostIdList(id);
+        vo.setPostIdList(postIdList);
+
+        return Result.ok(vo);
+    }
+
+    @GetMapping("info")
+    @Operation(summary = "登录用户信息")
+    public Result<SysUserVO> info() {
+        SysUserVO user = SysUserConvert.INSTANCE.convert(SecurityUser.getUser());
+
+        return Result.ok(user);
+    }
+
+    @PutMapping("password")
+    @Operation(summary = "修改密码")
+    public Result<String> password(@RequestBody @Valid SysUserPasswordVO vo) {
+        // 原密码不正确
+        UserDetail user = SecurityUser.getUser();
+        if (!passwordEncoder.matches(vo.getPassword(), user.getPassword())) {
+            return Result.error("原密码不正确");
+        }
+
+        // 修改密码
+        sysUserService.updatePassword(user.getId(), passwordEncoder.encode(vo.getNewPassword()));
+
+        return Result.ok();
     }
 
     @PostMapping
     @Operation(summary = "保存" )
     @PreAuthorize("hasAuthority('sys:user:save')" )
     public Result<String> save(@RequestBody SysUserVO vo) {
+        // 新增密码不能为空
+        if (StrUtil.isBlank(vo.getPassword())) {
+            Result.error("密码不能为空");
+        }
+
+        // 密码加密
+        vo.setPassword(passwordEncoder.encode(vo.getPassword()));
+
+        // 保存
         sysUserService.save(vo);
 
         return Result.ok();
@@ -60,6 +112,13 @@ public class SysUserController {
     @Operation(summary = "修改" )
     @PreAuthorize("hasAuthority('sys:user:update')" )
     public Result<String> update(@RequestBody @Valid SysUserVO vo) {
+        // 如果密码不为空，则进行加密处理
+        if (StrUtil.isBlank(vo.getPassword())) {
+            vo.setPassword(null);
+        } else {
+            vo.setPassword(passwordEncoder.encode(vo.getPassword()));
+        }
+
         sysUserService.update(vo);
 
         return Result.ok();
@@ -69,8 +128,33 @@ public class SysUserController {
     @Operation(summary = "删除" )
     @PreAuthorize("hasAuthority('sys:user:delete')" )
     public Result<String> delete(@RequestBody List<Long> idList) {
+        Long userId = SecurityUser.getUserId();
+        if (idList.contains(userId)) {
+            return Result.error("不能删除当前登录用户");
+        }
+
         sysUserService.delete(idList);
 
         return Result.ok();
+    }
+
+
+    @PostMapping("import")
+    @Operation(summary = "导入用户")
+    @PreAuthorize("hasAuthority('sys:user:import')")
+    public Result<String> importExcel(@RequestParam("file") MultipartFile file) {
+        if (file.isEmpty()) {
+            return Result.error("请选择需要上传的文件");
+        }
+        sysUserService.importByExcel(file, passwordEncoder.encode("123456"));
+
+        return Result.ok();
+    }
+
+    @GetMapping("export")
+    @Operation(summary = "导出用户")
+    @PreAuthorize("hasAuthority('sys:user:export')")
+    public void export() {
+        sysUserService.export();
     }
 }
